Add initial GV app; separate build tasks
Initial "app" is just Hello World, but it builds to public/gv2 as
a full React app, importing React through the npm package.

GeniBlocks examples still build to public/examples, using the standalone
GeniBlocks build product. This standalone product is still also built
to dist. (Geniverse app is not built to dist, as there is nothing to
distribute.)

GeniBlocks CSS, and a single vendor library for now (biologica.js) are
also copied to the public/gv2 folder. Any resources in src are copied
to gv2 (but not GeniBlocks, which should hopefully not need static
resources)