Rename all actions to be in past tense
https://github.com/reactjs/redux/issues/384

There is not a consensus on this issue, but it seemed right. "Actions"
could be better thought of as events, they describe a change that has
already occurred. The job of the reducer is to make the state reflect
that change.

Action creators (the functions below) are still written in the imperative.