Add navigate action that automatically changes route
This adds our own version of the router middleware, as the one provided
by the library caught the actions, to prevent logging, and also expected
its own actions, which weren't as semantically-meaningful.

Will add a proper version of `navigateNextChallenge` next, which will
require adding redux-thunk so we can look at state.