Load authored challenge automatically on navigation
Here we move `loadStateFromAuthoring` into the `NAVIGATED` reducer, which
means we will always load the current challenge whenever the user
navigates using an action. 

To support this, we capture the new case and challenge on navigation,
and also add the authoring directly to the state at the start (this
isn't the final version of how authoring will enter the state).

We can now eliminate the action `loadAuthoredChallenge`, and also
eliminate the lifecycle code in challenge-container to not need
to call loadAuthoredChallenge, as this will have already happened.

Note that this breaks navigating by url, which will require re-adding
those lifecycle methods.