Add modal dialog reducer; new actions, state, tests
This puts all the dialog information into a nested property in the state,
`modalDialog`, and requires an explicit action to make a dialog appear.

Dialogs no longer appear as side-effects to other actions, they must be
dispatched explicitly. This can be done from a single action creator using
redux-thunk to dispatch multiple actions at once (e.g. submit drake, then
show dialog).

NOTE: This breaks all current dialogs, and a number of tests, but they
will fixed on the next commit.

This also exposes all the properties of a dialog in the action, and gets
rid of the default strings being used in the modal-message-container.
These strings will need to be part of the action.

Add tests.