Add userHiddenDrake reducer
Added this in the middle of adding submitDrake message handling, because
we were originally setting the userHiddenDrake as a side effect as part
of the combined submitDrake/showMessage, and then when dismissing the
message.

Right now the drake is always hidden unless we are displaying a message,
because we only use this property in the genomeChallenge template.
If we need more context, we could always get the current template from the
state, and do more involved logic.