Add submitDrake dialog messages
Explicitly select the appropriate message to show after submitting a
drake.

This replaces logic that was contained both in reducer (setting
challengeCompleted) and in the message-container (finding out if
it was the last case) and removes that code from both.

Eventually we may still want to have a library of common messages to
draw from so not everything needs to be spelled out in the action
creators, but that's just for prettiness.