Notification container, reducer; wire-up ITS messages
This adds a new property to the state, an array of `notifications`.

Currently the simplistic behavior is that new ITS messages add to the
list of visible notifications, and all user actions clear the list.

Eventually we may want to have more complex behavior, like user-closing
and timed-closing messages, possibly defined from ITS.