Add changeable, visible genes to author, state, templates
This allows the authored properties `userChangeableGenes` and
`visibleGenes` to be accessed and used by all the templates.

For templates where alleles are always non-editable (egg challenges),
we still accept both, but the genome view will only show non-changeable
labels. (In theory even the egg games could allow certain alleles to
be changed, if the author chooses, but the chromosome-selection UI
becomes wonky if we do this, so I left it for now.)

Update the authoring to use the new properties.