Demo of authored limited-choice allele selection
This works with a new feature in biologica.js 0.0.3, which now supports
allele strings of the form

`a:x/y,b:x/y/z`

where x, y and z are options for alleles, and a random choice is made
among these. So, for example, if we do not want students to be
introduced to the Tk allele, we can author

`a:T/t,b:t/t`

Because of the way the dash-allele parser works, no changes need to be
made to GV 2.0 authoring, as `T/t-T/t` will be parsed as the above
automatically.

(Note, `bower install` does not always seem to bring in `latest`, so
`bower uninstall biologica.js` may be needed first.)