Add ability to randomizeTrials
When we first enter a challenge with multiple trials, we generate a
trial order, which in the default case will be [0...n].

When selecting the drakes to display, we pass `trialOrder[trial]` to
the template, instead of `trial`.

If `randomizeTrials: true` is in the authoring, then this array will
be shuffled, which will result in a random order of trials.