Initial round of text adjustments for Windows browsers. [#141964899]
All Windows browsers tested (Chrome, Firefox, IE11) showed the same text offsets, suggesting that it may be a font-metrics issue rather than a browser issue. We "fix" it for the time being by user agent-sniffing for Windows and adding the `os-windows` class to the `<body>` tag if it's present. Then CSS is used on a case-by-case basis to adjust the text where appropriate.