Add action for winning zoom challenge
The ZoomChallenge waits to hear a "challengeWon" event from the embedded
iframe through the iframe-phone, and when it hears it, calls the
ZOOM_CHALLENGE_WON action, which causes a crystal message to be shown.

This also adds a new ITS meta message, "USER COMPLETED CHALLENGE",
since none of the others seemed applicable.

[#144720089][#146319221]