Clean up fertilize action
This doesn't belong in this PR, but it's a "while I'm here" change.

The fertilize reducer doesn't use the gamete numbers, because it just
assumes it's using `state.currentGametes` 0 and 1. Not sure if this
was a change, or it's always been that way.

This cleans up the action to make that explicit.

This and the `breedClutch` action both assume they know the location of
the mother and father drakes, and they assume they know where to put
the offspring. In the case of `fertilize`, the offspring is always
placed in `drakes[2]`. Both of these assumptions might eventually need
to be changed, so the actions could/should have indices pointing to
the mother and father. Until we need it, though, we don't need to change
it.