Add currentGem prop, reducer; use it to show gem
This adds a new state property, currentGem, which currently just looks
at moves to set itself, like the "progress". It shows the gem a user
would receive if they finished the challenge immediately.

When the egg game is refactored, this will also look at score.

In this commit, progress is still being used to display old gems. This
will be shortly removed.