Replace challengeProgress with gems array
gems is a nested array in the state that shows the current gem awarded
for every challenge.

It is set by a reducer on two actions, CHALLENGE_COMPLETED (renamed to
make it consistent with other actions) and CHALLENGE_RETRIED (new action
to show when user hits retry).

The challenge-progress helper is kept for the time being to make it
easier for a view to get the array of gems for a mission, and to get
the score equivalencies (i.e, 3+ == no gem). This way that logic is in
one place. But we could also move this.

No test is in place for setting the gem array yet.