Show room when we enter a challenge
This updates loadStateFromAuthoring to add a new prop, showingRoom,
which is true when trial === 0. This may not be the best solution.

When showingRoom is true, we don't show the template, and change the css
classes to show the room background.

This also updates the bottom-hud to hide the trial counter, as well as
the scores and goalMoves, while we're showing the room.

Note, this commit does not have a way to get to the challenge yet.