Create notifications for dialog; add sample dialog
This allows us to add an array of dialog to the metadata of a challenge.

The array will simply be added to the notifications.messages object,
which will display them as it already does.

Instead of an array of strings, we allow an array of objects, so as to
include a `character` (not yet used), and also a `type`, which is used
to make decisions about the close button. This still supports plain
strings, but we could eventually wrap the ITS messages in objects as
well.

This hides the close button until the last line of dialog, and then
makes the close button trigger the ENTER action. The user can currently
also click on the hotspot at any time.