Navigation can show current level based on gems
If we open up the Navigation without a routeSpec (e.g. from `#home`),
then we use our current gems to figure out which level we are currently
on, and show that one.

The `getCurrentMissionFromGems` function will also be useful for
determining the dialog to show, which is why it returns other details.