Navigate back to home after last chal of mission
To navigate to home programmatically (instead of just typing #home in
the url) we have to manually update the route or we'll just get
re-routed to whatever challenge we were on, because the old route will
still be there (also the url should reflect the state).

Doing that means we need to avoid getting into an infinite loop when
ChallengeContainerSelector notices the route changes. So we have to
let it know the currentRoom.

This is very similar all around to what we are doing with the routes,
but it seemed like more of a hack to get the routeSpec to support the
home room, vs some duplication as it is now.