Continue button on victory will show map
If there are more challenges available in a mission, then the Continue
button will open up the Map. If there aren't, we will move to Home.

This is mediated by a new action, `continueFromVictory`.

Also clean up `navigateToNextChallenge` actions which are never never
used by containers, and fix showMap action to show map even if we are
currently showing victory.