Change gem state to be an array of historical gems
Instead of just saving/replacing the last gem earned for a challenge, we
now store all gems achieved, and append new gems to the end of the array.

`gems[level][mission][challenge]: 0`

is now

`gems[level][mission][challenge]: [0]`

This will allow the dashboard to show all the gems a student has earned.