Properly test completeChallenge, with new gem arr
We were skipping the submitDrake action because it calls a bunch of
thunks. This works through the logic of all the thunks called, in both
the nextTrial and the last trial state. It checks the gems, and that
the correct dialogs are shown in both situations.

The new gem array introduced by fe6c869 is tested when we check the
gem state.