Move stateVersion into saved state object
Instead of storing
```
{
  state: < savable state >,
  stateVersion: n,
  otherStuff: < other stuff >
}
```
which would require us to either make two calls to get the state and
version, or to get everything including `otherStuff`, we now store
stateVersion in state:
```
{
  state: {
    < savable state >,
    stateVersion: n
  },
  otherStuff: < other stuff >
}
```
When we load a state in, if we don't see a `stateVersion` we'll
assume it's version 1.

Simplify getUserQueryString since we are only ever requesting
`state` now.