Add migration system; migrate gems to v2; +tests
This adds a new migration system for migrating data, and migrates the
data before we call LOAD_SAVED_STATE.

(We also now call LOAD_SAVED_STATE with the entire state object, in case
any other data will be saved in state.)

Bump the stateVersion to 2, and add a migration to convert the old gems
format to the new format.

Get the stateVersion from `migrations` (single source of truth).

Add tests for migrations