Work backwards to find next challenge
Without gating, there may be holes where a user has skipped an
entire mission. In this case, with the old algorithm we would always
count the users "next" challenge as being the start of a new mission.
So it would always show the home-screen mission dialog between every
challenge.

This fixes it so that a user's current mission is either the last one
they have any gems for, or the next mission after that. It essentially
treats all the "holes" as being filled.

If a user completes 1.1.1, it will show the mission-start dialog for
2.1, and show the next challenge as being 2.1.1

If they next complete 2.2.1, it will show the next challenge as
being 2.2.2 (next challenge in 2.2)

If they next complete 2.3.2, it will show the next challenge as being
2.3.1 (first unstarted challenge in 2.3).

If the next finish all the challenges in 2.3, it will show the mission-
start dialog for 2.4, and show the next challenge as being 2.4.1.

[#150683738]