Make ITS action dispatch pure JSON
Previously we were dispatching an action with a GuideProtocol object
as the payload. This wasn't valid Redux and meant we couldn't fake
dispatches on the command line.

This moves the creation of the object to the reducer, so that we can
pass a plain JSON object as the payload.

We can now test tutor message from the Redux extension dispatcher with
an action such as

```
{
   type: "Guide message received",
   data: "{\"message\":{\"id\":\"ITS.CONCEPT.FEEDBACK\",\"text\":\"Let's look only at arms. What can you change to make a armless drake?\",\"args\":{\"trait\":\"forelimbs\"}},\"time\":1505901685372,\"reason\":{\"why\":\"MisconceptionDetected\",\"ruleId\":14}}"
}
```