Pass ITS trait to message; show highlight, arrow
If there is a `trait` property in the ITS args, we add that to the
message state.

In the view, we add a new div with that trait's class. Using CSS we
position the highlight in the right place, with an arrow image next
to it.

The CSS ensures that the highlight is shown only in layout-a, as other
layouts have not been tackled yet. It may turn out that other layouts
require something under more direct control from the template.

Add test.