Fix currentChallenge, don't pass level as argument
Previously we were passing in a level number when getting the current
challenge, but this doesn't match our current assumption, which is
that there is one single "next" challenge you are on.

This fixes the issue that every level-screen of the gems/challenge list
showed one highlighted gem.

This also fixes an off-by-one error caused when passing in the last
level.

[#151440205]