ClutchGame template supports submit-parents chall
This overloads the Clutch Game template to support the submit parents
for siblings challenges. The template and logic are very similar between
this and the create-unique challenge, so it makes sense to reuse.

The only major difference is the Submit Parents button, which calls an
action which checks if the current parents could be parents to all the
target drakes. If they can, then the call the submitParents action
with success=true.