Add randomMatched variation for trials; +test
In place of a single drake definition in a trial in authoring, we can
add an object with the property `randomMatched`, with an array of
possible drakes. When the trial is loaded, one of these variations will
be picked, the others will be ignored (they won’t be used in other
trials). The "matched" part of "randomMatched" means that we will pick
the same random index for all other drake definitions, after having
picked the first one randomly.

We can define the goal moves for that trial either as the regular integer,
or with an array. If the latter, we will use that same random index.

Add an example authoring document with randomMatched challenges.