Make moves counter count down; add gem colors
The moves counter now counts down instead of up, starting with goal
moves + 2, essentially showing the number of moves you can make and
still get a gem.

The color of the moves screen matches the gem you'll get if you make a
correct submission right then.

The logic for showing the moves counter remains, it's only on templates
that used to show moves.

[#146888383]