After the user completes challenge, show dialog
Instead of going directly to the Venture Map modal dialog on challenge
completion, where the user will be awarded a gem and see dialog, we
now show a default message in the dialog scroll while the challenge is
still showing first.

The gem is awarded immediately, so that the sound is heard and it
appears in the bar at the bottom, and the user's state will be saved
at that point, regardless of whether they move on.

Whether they close the dialog or hit the big green "Continue" button,
they are then taken to the Venture Map for the regular character
dialog.

[#154811093]