Rename challengeType "create-unique" to "match-target" in ClutchGame schema
ClutchGame uses three different challenge types, "submit-parents", "test-cross", and a third type that is not referenced directly in code, but is the default if the challengeType is not one of the other two. The value for that third type was previously "create-unique", but these are clearly target matching challenges for which the "match-target" type is clearly more appropriate. Since the value is never referenced in code, this change can be made entirely in the schema and the authoring document plus one CSS change.