Drop ITS messages as stale if they don't echo correct sequence id
This uses the currently unused `sequence` property of the ITS messages.

We update our `sequence` on each message sent, as we were already doing.
(The internal variable has been renamed to `lastActionSequenceId` to
make this clear.)

The ITS is now expected to echo back the sequence number when it sends a
message. If the ITS's sequence number is different, we will drop the
message as stale.

Currently, we assume that there is no case where we send the ITS two
messages, and we would want to react to a response to the first message.
Therefore, any such message is considered stale.

This is fault-tolerant if the ITS does not implement the sequence-echoing
protocol, or turns it off: if we get the default value of `-1`, as we
have currently been getting, we will not drop the message.

[#159735419][#159288382]