Add traits hint box to breeding challenges
This updates the backgroundClass code to allow templates to add additional
classes based on the state. In this case, the clutch game (layout H) will
add `fixed-mother` or `fixed-father` as appropriate.

The hint-arrow classes then use these to display on the left or the right.

We do not plan on showing hint boxes when both parents are changeable.

This works for match-drake and submit-parents clutch games.

[#158373203]