Egg game move counter decremented multiple times
The egg game decremented the move counter when an incorrect drake was submitted, and then again when the game was reset after an incorrect answer. Resetting the game should not decrement the move counter. This was also causing moves to decrement after a gamete was selected on two consecutive levels.

Furthermore, hatchSoundPlayed was not getting reset between consecutive Egg Game levels.